pyproject.toml
[project]
name = "product-comparison"
version = "0.1.0"
description = "Product comparison tool that scrapes e-commerce platforms and generates AI-powered comparisons"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31",
    "openai>=1.75.0",
    "google-genai>=1.70.0",
    "beautifulsoup4>=4.12",
    "lxml>=5.3",
    "python-dotenv>=1.0",
    "rich>=13.7",
    "fastapi>=0.135.3",
    "uvicorn[standard]>=0.44.0",
    "passlib[bcrypt]>=1.7.4",
    "pyjwt>=2.12.1",
    "bcrypt<4",
]

[project.scripts]
compare = "main:main"
